requirements.txt
torch>=2.0
torchvision>=0.15
numpy<2
pillow>=9.5
scikit-learn>=1.2
matplotlib>=3.7
tqdm>=4.65
pyyaml>=6.0
